+++ CHANGELOG.md
@@ -3,6 +3,15 @@ Changelog
 
 See also the [GitHub releases page](https://github.com/FriendsOfSymfony/FOSHttpCache/releases).
 
+2.4.0
+-----
+
+### Symfony
+
+* Introduced a new `CleanupCacheTagsListener` that allows removing the cache
+  tags header from the final response when sent to the client.
+
+
 2.3.1
 -----